# [ hnrm2 ] Use precision-enhanced hnrm2