Along with #2555 I inspected through every custom-made half-precision calculation function, and `hnrm2` is the only one left that needs f16-f32 precision supplement.


**Proposed Changes**
- Previous hnrm2 was using full-fp16.
- Since this is also one of dimension-shrinking computation, should use inter-fp32 values to enhance precision.
- This has not been detected due to small dimension Tensor usage in unittest. Add higher dimension test case accordingly.
- Note that this function is responsible for `Tensor::l2norm()`, frequently used for mse loss computation.

### Differnece w.r.t. fp32-cblas
| dim | f16 (prev) | f16f32 (now) |
| --- | --- | --- |
| 768 | 282.233 | 0.232574 |

### Latency
> mean value ( TC = 100)
> Since this is nanosec-unit result, almost trivial

| dim | f16 | f16f32 | fp32-cblas | 
| --- | --- | --- | --- |
| 768 | 508 ns | 579 ns | 4242 ns |

Better accuracy with almost no latency deterioration

**Self evaluation:**
1. Build test:     [X]Passed [ ]Failed [ ]Skipped
2. Run test:     [X]Passed [ ]Failed [ ]Skipped